% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datlist.R
\docType{data}
\name{datlist}
\alias{datlist}
\title{Simulated trial data}
\format{
An object of class \code{"list"}
\describe{
\item{subjid}{Subject ID from 1 to maximum sample size}
\item{arm}{Treatment arm allocation for each subject}
\item{dat}{Response for each subject (0 or 1)}
\item{arrival.day}{Arrival time (days) for each subject throughout trial duration}
\item{obstime}{Obveration time (days) for each subject; when response is observed}
}
}
\usage{
data(datlist)
}
\description{
Simulated trial data list \code{datlist} that is generated by \code{singleTrial.R} and contains
all the necessary information to perform confidence analysis.
}
\examples{

data(datlist)
head(datlist)

}
\references{
This data set was created for the confidenceSim package.
}
\keyword{datasets}
