% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialDesignFns.R
\name{getDataBin}
\alias{getDataBin}
\title{Generate binary data}
\usage{
getDataBin(arm, resprate)
}
\arguments{
\item{arm}{Arm allocation for a single patient.Expects number in \emph{1,2,...,n} where \emph{n} the number of treatment arms including control.}

\item{resprate}{Response rates for each arm. Expects a vector of probabilities of length \emph{n} with the first
corresponding to response rate of the control arm.}
}
\value{
Returns a binary value corresponding to patient response.
}
\description{
Given an arm allocation and response rates, this function generates a binary response.
}
\examples{

response <- getDataBin(1, c(0.5, 0.7))

}
