% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\docType{data}
\name{inputs}
\alias{inputs}
\title{Example trial input parameters}
\format{
An object of class \code{"list"}
\describe{ Parameter list used to generate a trial, in this case a 2-arm 6-stage trial with
options to stop early for efficacy, inferiority or futility based on confidence thresholds
}
}
\usage{
data(inputs)
}
\description{
List of parameter that contains
all the necessary information to generate a trial.
}
\examples{

data(inputs)
print(inputs)

}
\references{
This data set was created for the confidenceSim package.
}
\seealso{
\code{\link[=getparlist]{getparlist()}}
}
\keyword{datasets}
