% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialDesignFns.R
\name{getparlist}
\alias{getparlist}
\title{Get Parameter List}
\usage{
getparlist(
  looks = seq(500, 1000, 100),
  nmax = NULL,
  perpetual = FALSE,
  alloc.ratio = c(1, 1),
  num.per.block = c(1, 1),
  final.visit = 0,
  as.type = "asOF",
  alpha = 0.05,
  multiarm.mode = "CONFIDENCE-BASED",
  lmb = 0.1,
  lmb.conf.thresh = 0.9,
  outcome.type = "BINARY",
  estimator.type = "odds ratio",
  resprate = c(0.3, 0.5),
  ppm = rep(15, 300),
  special = NULL
)
}
\arguments{
\item{looks}{Vector of analysis times expressed by either number of patients accrued at each point or by rate of information accumulated.
If the former, last item should be the maximum sample size.
If the latter, last item is 1.
Expects a vector with length equal to the total number of total looks.}

\item{nmax}{Maximum sample size, specified if information rates are used for 'looks'.}

\item{perpetual}{Whether to run the trial perpetually (TRUE) or not (FALSE).
If TRUE, new treatment arms will be added when treatment arms are dropped, until there are no more arms left.
All treatments are included via the 'resprate' parameter. Default is FALSE.}

\item{alloc.ratio}{Allocation ratios for study arms relative to each other.
Expects vector with length equal to number of arms including control.
First number corresponds to control ratio.}

\item{num.per.block}{Number from each arm per block, for blocked randomization to balance co-variates.
Block size is 'sum(num.per.block)'. If a single number is provided, it will be assume to apply to each arm.}

\item{final.visit}{The number of days after intervention when the response information becomes available.}

\item{as.type}{The type of alpha spending function to use in group sequential design.
Default is 'asOF', O'Brien-Fleming-type.}

\item{alpha}{The alpha threshold to apply to each pairwise comparison to control in the final analysis.
Used together with the 'MONITOR FUTILITY', when an alpha spending function is not needed.
Default is 0.05, assuming a two-sided test.}

\item{multiarm.mode}{For multiple treatment arms, describes how arms are evaluated at each stage:
\itemize{
\item{"CONFIDENCE-BASED"(default): Evaluate arms against confidence-based rules }
\item{"DROP WORST": Drop the worst performing arm, and carry the remaining promising arms}
\item{"SELECT BEST": Select the best performing arm to carry forward, drop the rest}
\item{"ALL PROMISING": Carry forward all promising arms}
\item{"MONITOR FUTILITY": Only monitor for futility}
}}

\item{special}{}
}
\description{
Generate a parameter list to generate a frequentist confidence trial
}
